refactor: Update operation names for withdraw and withdraw to debt
