+++ packages/deploy-configurations/operation-definitions/index.ts
@@ -35,8 +35,8 @@ export { getSparkDepositOperationDefinition } from './spark/borrow'
 export { getSparkDepositBorrowOperationDefinition } from './spark/borrow'
 export { getSparkOpenDepositBorrowOperationDefinition } from './spark/borrow'
 export { getSparkPaybackWithdrawOperationDefinition } from './spark/borrow'
-export { getSparkWithdrawOperationDefinition } from './spark/borrow/withdraw'
-export { getSparkWithdrawToDebtOperationDefinition } from './spark/borrow/withdraw-to-debt'
+export { getSparkWithdrawOperationDefinition } from './spark/auto/withdraw'
+export { getSparkWithdrawToDebtOperationDefinition } from './spark/auto/withdraw-to-debt'
 export { getSparkOpenOperationDefinition } from './spark/multiply'
 export { getSparkCloseAndExitOperationDefinition } from './spark/multiply'
 export { getSparkCloseAndRemainOperationDefinition } from './spark/multiply'
